Update README.md
Introduction

Write a client program that connects to a server. The server hosts electrocardiogram (ECG) data points of 15 patients suffering from cardiac diseases. The client communicates with the server to complete two tasks:

Obtain individual data points from the server.

Obtain a whole raw file of any size in one or more segments from the server.

The client has to send properly-formatted messages to the server using a communication protocol defined by the server.

Starter Code

You are given a source directory with the following files:

makefile (makefile)

This file compiles and builds the source files when you type the make command in the terminal.

FIFORequestChannel class (FIFORequestChannel.cpp/.h)

These files implements a pipe-based communication channel. The client and server processes use this class to communicate with each other. This class has a read and a write function to receive and send data from/to another process, respectively. The usage of the function is demonstrated in the given client.cpp. Do not modify this class.